move observers and data accessor to there new packages (architecture meaning is better)
