+++ MyPenelopF/src/main/java/Observer/ContactListener.java
@@ -1,4 +1,4 @@
-package utils;
+package Observer;
 
 import classes.Contact;
